Add files via upload
This machine learning project focuses on predicting the quality of a particular type of wine, Red, using the Winequality-red.csv dataset. The dataset contains 11 attributes as well as the quality rating of the wine. Using supervised learning techniques such as regression, the goal of this project is to build a model to accurately predict the quality of the wine based on the attributes. The model will be evaluated using metrics such as accuracy, precision, and recall. The model will then be used to predict the quality of new wine samples. In addition, the project will explore feature selection and engineering techniques to improve the performance of the model.